--- a/library/src/methods/strict/strict.ts
+++ b/library/src/methods/strict/strict.ts
@@ -1,5 +1,4 @@
 import type { ObjectSchema } from '../../schemas/object/index.ts';
-import { getIssue } from '../../utils/index.ts';
 
 /**
  * Creates a strict object schema that throws an error if an input contains
@@ -33,12 +32,13 @@ export function strict<TSchema extends ObjectSchema<any>>(
           Object.keys(result.output).length
         ? {
             issues: [
-              getIssue(info, {
+              {
+                type: 'leaf',
                 reason: 'object',
                 validation: 'strict',
                 message: error || 'Invalid keys',
                 input,
-              }),
+              },
             ],
           }
         : result;