--- a/library/src/schemas/any/anyAsync.ts
+++ b/library/src/schemas/any/anyAsync.ts
@@ -1,5 +1,5 @@
 import type { BaseSchemaAsync, PipeAsync } from '../../types.ts';
-import { executePipeAsync, getPipeInfo } from '../../utils/index.ts';
+import { executePipeAsync } from '../../utils/index.ts';
 
 /**
  * Any schema type.
@@ -36,7 +36,7 @@ export function anyAsync(pipe: PipeAsync<any> = []): AnySchemaAsync {
      * @returns The parsed output.
      */
     async _parse(input, info) {
-      return executePipeAsync(input, pipe, getPipeInfo(info, 'any'));
+      return executePipeAsync(input, pipe, info, 'any');
     },
   };
 }