--- a/library/src/schemas/bigint/bigint.ts
+++ b/library/src/schemas/bigint/bigint.ts
@@ -1,5 +1,9 @@
 import type { BaseSchema, Pipe } from '../../types.ts';
-import { executePipe, getErrorAndPipe, getIssue } from '../../utils/index.ts';
+import {
+  executePipe,
+  getErrorAndPipe,
+  getLeafIssue,
+} from '../../utils/index.ts';
 
 /**
  * Bigint schema type.
@@ -59,7 +63,7 @@ export function bigint(
       if (typeof input !== 'bigint') {
         return {
           issues: [
-            getIssue(info, {
+            getLeafIssue({
               reason: 'type',
               validation: 'bigint',
               message: error || 'Invalid type',