--- a/library/src/schemas/blob/blobAsync.ts
+++ b/library/src/schemas/blob/blobAsync.ts
@@ -2,7 +2,7 @@ import type { BaseSchemaAsync, PipeAsync } from '../../types.ts';
 import {
   executePipeAsync,
   getErrorAndPipe,
-  getIssue,
+  getLeafIssue,
 } from '../../utils/index.ts';
 
 /**
@@ -66,7 +66,7 @@ export function blobAsync(
       if (!(input instanceof Blob)) {
         return {
           issues: [
-            getIssue(info, {
+            getLeafIssue({
               reason: 'type',
               validation: 'blob',
               message: error || 'Invalid type',