--- a/library/src/schemas/boolean/booleanAsync.ts
+++ b/library/src/schemas/boolean/booleanAsync.ts
@@ -3,7 +3,6 @@ import {
   executePipeAsync,
   getErrorAndPipe,
   getIssue,
-  getPipeInfo,
 } from '../../utils/index.ts';
 
 /**
@@ -81,7 +80,7 @@ export function booleanAsync(
       }
 
       // Execute pipe and return result
-      return executePipeAsync(input, pipe, getPipeInfo(info, 'boolean'));
+      return executePipeAsync(input, pipe, info, 'boolean');
     },
   };
 }