--- a/library/src/schemas/date/date.ts
+++ b/library/src/schemas/date/date.ts
@@ -1,10 +1,5 @@
 import type { BaseSchema, Pipe } from '../../types.ts';
-import {
-  executePipe,
-  getErrorAndPipe,
-  getIssue,
-  getPipeInfo,
-} from '../../utils/index.ts';
+import { executePipe, getErrorAndPipe, getIssue } from '../../utils/index.ts';
 
 /**
  * Date schema type.
@@ -75,7 +70,7 @@ export function date(
       }
 
       // Execute pipe and return result
-      return executePipe(input, pipe, getPipeInfo(info, 'date'));
+      return executePipe(input, pipe, info, 'date');
     },
   };
 }