--- a/library/src/schemas/date/date.ts
+++ b/library/src/schemas/date/date.ts
@@ -1,5 +1,9 @@
 import type { BaseSchema, Pipe } from '../../types.ts';
-import { executePipe, getErrorAndPipe, getIssue } from '../../utils/index.ts';
+import {
+  executePipe,
+  getErrorAndPipe,
+  getLeafIssue,
+} from '../../utils/index.ts';
 
 /**
  * Date schema type.
@@ -59,7 +63,7 @@ export function date(
       if (!(input instanceof Date)) {
         return {
           issues: [
-            getIssue(info, {
+            getLeafIssue({
               reason: 'type',
               validation: 'date',
               message: error || 'Invalid type',