--- a/library/src/schemas/nan/nan.ts
+++ b/library/src/schemas/nan/nan.ts
@@ -1,5 +1,5 @@
 import type { BaseSchema } from '../../types.ts';
-import { getIssue } from '../../utils/index.ts';
+import { getLeafIssue } from '../../utils/index.ts';
 
 /**
  * NaN schema type.
@@ -40,7 +40,7 @@ export function nan(error?: string): NanSchema {
       if (!Number.isNaN(input)) {
         return {
           issues: [
-            getIssue(info, {
+            getLeafIssue({
               reason: 'type',
               validation: 'nan',
               message: error || 'Invalid type',