--- a/library/src/utils/getPathInfo/index.ts
+++ b/library/src/utils/getPathInfo/index.ts
@@ -1,1 +1,0 @@
-export * from './getPathInfo.ts';