Hey,

Recently popped into [this issue thread](https://github.com/fabian-hiller/valibot/issues/73) to discuss a fairly significant perf improvement idea, but found out @fabian-hiller had already done almost exactly the same thing!

However, I just finished a 4 hour plane ride and I spent the whole flight looking for more perf improvements and found a few! 

**Sizeable optimizations**

Both of these generally just involved doing less whenever possible:

1. Evaluate full issue paths lazily instead of eagerly
2. Avoid creating PipeInfo objects when `pipe.length === 0`

I'd say these contributed 90% to the total performance improvement in this PR.

**Micro-optimizations**

1. Make `cachedEntries` for the `object` schema a flat array instead of an array of 2-tuples -- mainly just less indirection which is more friendly to the CPU.
2. Simpler "is object" check using `input?.constructor === Object`


Here are the results on my machine using [typescript-runtime-type-benchmarks]():

```
*** valibot @ current head of main, i.e. e6c53c8 ***
Running "parseSafe" suite...
Progress: 100%

  valibot:
    1 784 335 ops/s, ±1.51%   | fastest

Finished 1 case!
Running "parseStrict" suite...
Progress: 100%

  valibot:
    1 715 007 ops/s, ±0.15%   | fastest

-------------------------------------------

*** THIS PR ***
Running "parseSafe" suite...
Progress: 100%

  valibot:
    2 616 009 ops/s, ±0.21%   | fastest

Finished 1 case!
Running "parseStrict" suite...
Progress: 100%

  valibot:
    2 317 246 ops/s, ±0.25%   | fastest
```

This run amounts to +47% on `parseSafe`, +35% on `parseStrict`.