@fabian-hiller 2023-08-15
Thank you very much for your contribution! At the moment I have to focus on the text of my bachelor thesis. As soon as I have more time, I will check and merge your changes.