@milankinen 2023-08-16
Just a quick notice about path if we're really trying to optimize the performance. Usually path is only needed in case of failures so that the path can be added to the issues. If this is the case, it can be allocated lazily only if the validation (or sub-validation) returns any issues instead of the [current implementation where the path is allocated eagerly](https://github.com/jussisaurio/valibot/blob/perf-improvements/library/src/schemas/object/object.ts#L136-L141). This should yield some small performance improvements due to decreased number of allocations and gc overhead. 

If the failure types are something like 
```ts
type Issues = { 
  pathItem: PathItem
  issues: (IssueErrorMessageAndMeta | Issues[])[] 
}
````
Then the `_parse` function validation logic in failure cases becomes extremely simple:
```ts
// object validation pseudo
let issues: Issues[] | undefined
for each key in objectSchema {
  const result = parseKeySchema(object[key])
  if (result.issues) {
   (issues = issues ?? []).push(result.issues)
  }
}
return issues ? { valid: false, issues } : { valid: true, value }
```

Of course the issues structure (and paths along with it) must be flattened at the top level parse before returning the issues to the library user, but that should be very straightforward task to do (similar to the current `getPath`).

That said, I'm not familiar with this library or its features so take this with grain of salt... 😄