@fabian-hiller 2023-08-16
Yes, that is correct. Thank you very much for pointing it out. I will take it into account for the final implementation.