@jussisaurio 2023-08-16
Yep that's a very good point, interested to see how much this speeds things up further