@fabian-hiller 2023-08-17
Thank you for your contribution! Due to my bachelor thesis I have a lot to do at the moment. I'll try to check the changes next week at the latest.