@jussisaurio 2023-08-17
> Thank you for your contribution! Due to my bachelor thesis I have a lot to do at the moment. I'll try to check the changes next week at the latest.

Sure, no pressure nor hurry, it's your library dude :) I just rebased and fixed the conflicts, nothing else.

If I have extra time I can try to find time to implement @milankinen 's suggested refactor as well, but it's kind of 50/50 whether I can find the time