@jussisaurio 2023-08-17
Found an hour of idle time and started shoveling, looking pretty promising... :D

```
Removing previous results
Executing "valibot"
Loading "valibot"
Running "parseSafe" suite...
Progress: 100%

  valibot:
    3 657 558 ops/s, ±0.11%   | fastest

Finished 1 case!
Running "parseStrict" suite...
Progress: 100%

  valibot:
    3 213 696 ops/s, ±0.10%   | fastest
```

@milankinen you absolute GOAT! 🐐 

I'm not quite finished with the refactor yet - must make sure all the tests pass and none of the public API is broken.