@fabian-hiller 2023-08-17
Can you explain why it is necessary to make a distinction between `LeafIssue`, `NestedIssue` and `UnionIssue` for the issues?