@jussisaurio 2023-08-17
> Can you explain why it is necessary to make a distinction between `LeafIssue`, `NestedIssue` and `UnionIssue` for the issues?

Sure! We can obviously iterate on the data model, but the idea in a nutshell is to accurately model the different kinds of objects according to the information that is actually relevant to each:

`LeafIssue` is an issue with a "primitive": the input was supposed to be a string, but was a number instead, or was supposed to be an object, but was a string instead (important to distinguish this from other issues with objects -- here the specific problem is that the input is _not an object at all_). So for these types of issues it makes sense to include `input` , `message` etc, but it does not include other issues, so they shouldn't be in its datatype. The primitive also doesn't really have to be aware of `path` -- a string is a string, the path is its containing parent's concern.

---

`NestedIssue` is an issue with a "composite": an input may be a valid object (i.e. it is not a string), but it can have e.g. the following problem: it should have contained key `a` of `string`, but `a` was a `number` instead. It's still a valid object, but something within its composition is wrong. More precisely, one or more sub-`path`s within this composite contained issues. Hence, it makes sense to include `path` and a list of `issues` from each subpath. Likewise, for maps and records, `origin` belongs here, but nowhere else, since `origin` is nearly always `value`. 

- Looking at the current implementation of `NestedIssue`, it should probably still have the `validation` property which I neglected to add. But does it need to have `reason`? The reason in the above example is that one of its child properties was not a string. So I'm thinking the `reason` belongs to the child, not the container.

---

`UnionIssue` was something that I identified as a special case: It's basically a "pathless" NestedIssue. Not 100% sure what to do with it. The current data model for it may well not be the right one.