@milankinen 2023-08-18
Wow nicely done @jussisaurio 😮 👌!

>Can you explain why it is necessary to make a distinction between LeafIssue, NestedIssue and UnionIssue for the issues?

Sorry my fault, separation was just my initial thought because of separation of concerns. 😅  It's not absolutely necessary to have different types of issues if the issues are flattened "in-place" in schemas. E.g. something like:

```ts
const createIssue = (msg: string) => ({msg, path: []})
const appendIssues = (issues: Issue[], childIssues: Issue[], pathItem: PathItem) => {
  issues.push(...childIssues.map(issue => issue.path.unshift(pathItem) && issue))
  return issues
}

// string.ts
if (typeof value !== 'string') {
  return createIssue('not a string')
}

// object.ts
if (result.issues) {
  issues = appendIssues(issues, result.issues, key)
}
```