@fabian-hiller 2023-08-18
I don't have time to think about it in detail at the moment. However, I want to keep the library as simple as possible, as this also reduces the bundle size, and splitting it into different issues seems a bit complicated. It is already possible to extract all important information with the current implementation, isn't it?

A path is still needed for a `LeafIssue` if it is in an object. Or can a `LeafIssue` not concern nested data? If so, the name is not optimal.

The only thing that would make sense to me would be:
- `RootIssue`: Issue without path
- `NestedIssue`: Issue with path
- `UnionIssue` (special case): With sub-issues (stems from `RootIssue` or `NestedIssue`)

With the current implementation, the path is only constructed once an issue occurs, which is great. Does this lead to bugs with deeply nested issues? We may construct the path retroactively e.g. with `path.unshift(…)`? Then `LazyPath` would not be necessary, and the code would be simpler constructed or?