@fabian-hiller 2023-08-18
Tomorrow or Sunday, I will try to release a new version with all performance improvements. For time reasons (because of my bachelor thesis), I will then weigh whether I build on this issue or implement the ideas independently for the time being. I appreciate your contribution and will mention you in the release notes. 🤝