@jussisaurio 2023-08-18
@fabian-hiller you are free to do what you like ofc :) the important point is the lazy construction of paths.

as you can see in the most recent commit i removed the LazyPath class entirely because its not needed at all. ParseInfo no longer contains path because it's only constructed when an issue actually arises. that is the key detail.

wish you the best with your thesis, those are a lot of work!